Implement logic for handing split hands
The player can now choose whether to have a card dealt to a hand or
stand that hand. Once a hand is finished through stand, 21 or bust, the
game moves onto the next hand, or continues the game if it was the
player's second hand. (Still need to rewrite House's logic for split
hands)